Extract cloud CTA on search home
