+++ client/web/src/search/home/QueryExamplesHomepage.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useEffect } from 'react'
+import React, { useCallback, useState } from 'react'
 
 import { mdiOpenInNew } from '@mdi/js'
 import classNames from 'classnames'
@@ -8,12 +8,9 @@ import { SyntaxHighlightedSearchQuery } from '@sourcegraph/branded'
 import { AuthenticatedUser } from '@sourcegraph/shared/src/auth'
 import { EditorHint, QueryState } from '@sourcegraph/shared/src/search'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-import { buildCloudTrialURL } from '@sourcegraph/shared/src/util/url'
 import { Button, H2, Link, Icon, Tabs, TabList, TabPanels, TabPanel, Tab } from '@sourcegraph/wildcard'
 
-import { CloudCtaBanner } from '../../components/CloudCtaBanner'
-import { SearchPatternType } from '../../graphql-operations'
-import { eventLogger } from '../../tracking/eventLogger'
+import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 
 import { exampleQueryColumns } from './QueryExamplesHomepage.constants'
 import { useQueryExamples, QueryExamplesSection } from './useQueryExamples'
@@ -111,43 +108,10 @@ export const QueryExamplesHomepage: React.FunctionComponent<QueryExamplesHomepag
         ]
     )
 
-    const [cloudCtaVariant, setCloudCtaVariant] = useState<CloudCtaBanner['variant'] | string>('filled')
-    useEffect(() => {
-        const searchParams = new URL(window.location.href).searchParams
-        const uxParam = searchParams.get('cta')
-        const allowedVariants: { [key: string]: string | undefined } = {
-            a: 'filled',
-            b: 'underlined',
-            c: 'outlined',
-            d: undefined,
-        }
-
-        if (uxParam && Object.keys(allowedVariants).includes(uxParam)) {
-            setCloudCtaVariant(allowedVariants[uxParam])
-        }
-    }, [])
-
     return (
         <div>
             {isSourcegraphDotCom ? (
                 <>
-                    <div className="d-table mx-auto">
-                        <CloudCtaBanner className="mb-5" variant={cloudCtaVariant}>
-                            To search across your private repositories,{' '}
-                            <Link
-                                to={buildCloudTrialURL(authenticatedUser)}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                onClick={() =>
-                                    eventLogger.log('ClickedOnCloudCTA', { cloudCtaType: 'HomeUnderSearch' })
-                                }
-                            >
-                                try Sourcegraph Cloud
-                            </Link>
-                            .
-                        </CloudCtaBanner>
-                    </div>
-
                     <Tabs size="medium" onChange={handleTabChange}>
                         <TabList wrapperClassName={classNames('mb-4', styles.tabHeader)}>
                             <Tab key="Code search basics">Code search basics</Tab>
